refactor(helpers): extract render context from scene
- Move render context types and lifecycle helpers from scene helpers into a dedicated context module
- Re-export the new context helper module from the shared helpers entrypoint
- Split helper tests so context behavior is covered separately from scene accessors
