+++ __tests__/helpers/scene.test.ts
@@ -1,14 +1,6 @@
 import { Scene } from 'phaser';
 
-import {
-  createRenderContext,
-  getScene,
-  setScene,
-} from '../../src/helpers/scene';
-
-vi.mock('../../src/render/reconcile', () => ({
-  reconcileTree: vi.fn(() => null),
-}));
+import { getScene, setScene } from '../../src/helpers/scene';
 
 const scene = new Scene();
 
@@ -26,40 +18,5 @@ describe('setScene', () => {
   it('sets scene', () => {
     expect(setScene(scene)).toEqual(undefined);
     expect(getScene()).toEqual(scene);
-  });
-});
-
-describe('createRenderContext', () => {
-  it('resetStateIndex resets state index to 0', () => {
-    const context = createRenderContext();
-    context.getNextStateIndex();
-    context.getNextStateIndex();
-    context.resetStateIndex();
-    expect(context.getNextStateIndex()).toBe(0);
-  });
-
-  it('rerender calls componentFn when componentFn, componentProps, and scene are set', () => {
-    const componentFn = vi.fn((): null => null) as unknown as Parameters<
-      typeof createRenderContext
-    >[2];
-    const componentProps = { x: 1 };
-    const context = createRenderContext(
-      null,
-      scene,
-      componentFn,
-      componentProps,
-    );
-    context.rerender();
-    expect(componentFn).toHaveBeenCalledWith(componentProps);
-  });
-
-  it('rerender calls reconcileTree with element when only element and scene are set', async () => {
-    const { reconcileTree } = await import('../../src/render/reconcile');
-    const element = { type: () => null, props: {} } as unknown as Parameters<
-      typeof createRenderContext
-    >[0];
-    const context = createRenderContext(element, scene);
-    context.rerender();
-    expect(reconcileTree).toHaveBeenCalledWith(element, null, scene);
   });
 });
